# obfuscate: clean up SQL obfuscator