--- a/obfuscate/sql_test.go
+++ b/obfuscate/sql_test.go
@@ -81,7 +81,6 @@ func TestSQLResourceWithError(t *testing.T) {
 
 		NewObfuscator(nil).Obfuscate(&tc.span)
 		assert.Equal("Non-parsable SQL query", tc.span.Resource)
-		assert.Equal("Query not parsed", tc.span.Meta["agent.parse.error"])
 		assert.Equal(testSpan.Resource, tc.span.Meta["sql.query"])
 	}
 }
@@ -374,10 +373,10 @@ func TestMultipleProcess(t *testing.T) {
 	}
 
 	// The consumer is the same between executions
-	consumer := newTokenConsumer()
+	obf := newSQLObfuscator()
 
 	for _, tc := range testCases {
-		output, err := consumer.Process(tc.query)
+		output, err := obf.obfuscate(tc.query)
 		assert.Nil(err)
 		assert.Equal(tc.expected, output)
 	}
@@ -389,9 +388,9 @@ func TestConsumerError(t *testing.T) {
 	// Malformed SQL is not accepted and the outer component knows
 	// what to do with malformed SQL
 	input := "SELECT * FROM users WHERE users.id = '1 AND users.name = 'dog'"
-	consumer := newTokenConsumer()
+	obf := newSQLObfuscator()
 
-	output, err := consumer.Process(input)
+	output, err := obf.obfuscate(input)
 	assert.NotNil(err)
 	assert.Equal("", output)
 }
@@ -405,14 +404,14 @@ func BenchmarkTokenizer(b *testing.B) {
 		{"Escaping", `INSERT INTO delayed_jobs (attempts, created_at, failed_at, handler, last_error, locked_at, locked_by, priority, queue, run_at, updated_at) VALUES (0, '2016-12-04 17:09:59', NULL, '--- !ruby/object:Delayed::PerformableMethod\nobject: !ruby/object:Item\n  store:\n  - a simple string\n  - an \'escaped \' string\n  - another \'escaped\' string\n  - 42\n  string: a string with many \\\\\'escapes\\\\\'\nmethod_name: :show_store\nargs: []\n', NULL, NULL, NULL, 0, NULL, '2016-12-04 17:09:59', '2016-12-04 17:09:59')`},
 		{"Grouping", `INSERT INTO delayed_jobs (created_at, failed_at, handler) VALUES (0, '2016-12-04 17:09:59', NULL), (0, '2016-12-04 17:09:59', NULL), (0, '2016-12-04 17:09:59', NULL), (0, '2016-12-04 17:09:59', NULL)`},
 	}
-	consumer := newTokenConsumer()
+	obf := newSQLObfuscator()
 
 	for _, bm := range benchmarks {
 		b.Run(bm.name+"/"+strconv.Itoa(len(bm.query)), func(b *testing.B) {
 			b.ResetTimer()
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
-				_, _ = consumer.Process(bm.query)
+				_, _ = obf.obfuscate(bm.query)
 			}
 		})
 	}